fix(home): complete minor functionality for home pages

+++ rental-service/src/main/java/com/bikalp/rentalservice/entity/Room.java
@@ -43,7 +43,7 @@ public class Room {
 
     private String address;
 
-    private double numberOfRooms;
+    private int numberOfRooms;
 
     private int capacity; // number of people
